Continuously Adaptive Meanshift Tracking in Videos
CamShift is a tracking algorithm based on dynamic color probability distribution such that the distributions could be recomputed for every frame. The algorithm anticipates the movement of the object and quickly tracks it in the next frame of the video. It is based on tracking the hue of the object and frames will be converted to HSV color space. Histogram back projection method is used to achieve that.